# src/Presentation/SampleApi/Swagger/v3/swagger.json
{
  "x-generator": "NSwag v14.1.0.0 (NJsonSchema v11.0.2.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "Clean Architecture template",
    "description": "A template based on Clean Architecture rules. This template aims to show/help developers with setting up a solution based on Clean Architecture rules and guidelines.",
    "version": "v3"
  },
  "paths": {
    "/v3/player/new": {
      "post": {
        "tags": [
          "Player"
        ],
        "summary": "Create new Player\nPOST is equivalent to CREATE in CRUD.",
        "operationId": "CreateNewPlayerAsync",
        "requestBody": {
          "x-name": "requestModel",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlayerRequestModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "The value of the primary key",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true,
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Result<Int32>"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Result<Int32>": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/ResultStatus"
          },
          "error-messages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "validation-errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          },
          "success-message": {
            "type": "string"
          }
        }
      },
      "ResultStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "None",
          "Ok",
          "Error",
          "NotFound",
          "Invalid"
        ],
        "enum": [
          "none",
          "ok",
          "error",
          "not-found",
          "invalid"
        ]
      },
      "ValidationError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error-message": {
            "type": "string",
            "nullable": true
          },
          "error-code": {
            "type": "string",
            "nullable": true
          },
          "validation-error-type": {
            "$ref": "#/components/schemas/ValidationErrorType"
          }
        }
      },
      "ValidationErrorType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "None",
          "Error",
          "Warning",
          "Info"
        ],
        "enum": [
          "none",
          "error",
          "warning",
          "info"
        ]
      },
      "CreatePlayerRequestModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "first-name": {
            "type": "string"
          },
          "lastname": {
            "type": "string",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Player",
      "description": "Sample Controller for the Clean Architecture proposal"
    }
  ]
}